refactor sorting orbits into a hook
